requirement.txt
boto3
requests
google-cloud-storage==2.9.0